ThreadPool!!
Big day for the future employed (a play on a tiktok reference)

Worked on getting better documentation for methods like hangle_connect. made the error produced by read_to_string print to error output and return a 500 Internal server errror page instead of panicking! good improvment IMO

The most imporant part was getting the threadpool set up as well as getting the related structs and implementation blocks going.
This was hard and I had to rely on the book as a reference a lot.
Which makes sense because the book taught me how to do this. And thus looking back to it for help kinda makes my code look very similar. I tried to clean it up a bit and add more documentation, as well as more error handling as practcie though.