moved files, trying to make parser
The parser is done, but the issue is how im handling errors.

if an error happens because there isnt a file in public/ then im currenly passing the path into write_response, not the html body of 404 or any other response. so this needs to be fixed. I think an enum might be the way to go and then recalling read_file to get the correct body.