Played around with jetbrains auto documentation
I need to learn the proper way to document structs, enums, fn, methods, etc... So for this file I ran the Jetbrains auto documentation feature to see how idiomatic documentation looks and learn from it. I dont love using AI for this, but in this case I will learn from it and then write my own documentation for the next phases of this project.