Starting asnyc with mio!
Getting nonblocking.rs set up with a couple structs and enums.
next thing is that there are a couple errors with response.rs that need to be fixed. As well as a refactor to take writing out of response.rs and move that into nonblocking.rs and feed things into and out of the event loop to make things more modular.